+++ devtools/vite-env.d.ts
@@ -1,0 +1,11 @@
+/// <reference types="vite/client" />
+
+interface ImportMetaEnv {
+  VITE_AYAME_SIGNALING_URL: string
+  VITE_AYAME_ROOM_ID: string
+  VITE_AYAME_SIGNALING_KEY: string
+}
+
+interface ImportMeta {
+  readonly env: ImportMetaEnv
+}