+++ devtools/vite.config.mjs
@@ -1,0 +1,13 @@
+import { resolve } from 'node:path'
+import { defineConfig } from 'vite'
+
+export default defineConfig({
+  root: resolve(__dirname),
+  dist: resolve(__dirname, 'dist'),
+  resolve: {
+    alias: {
+      '@open-ayame/ayame-web-sdk': resolve(__dirname, '../dist/ayame.mjs'),
+    },
+  },
+  envDir: resolve(__dirname, '..'),
+})