+++ src/utils.ts
@@ -1,5 +1,3 @@
-import type { VideoCodecOption } from './connection/options'
-
 /**
  * @ignore
  */
@@ -13,20 +11,7 @@ interface WindowPerformance {
 declare let window: Window
 
 /**
- * @ignore
- */
-export function randomString(strLength: number): string {
-  const result = []
-  const charSet = '0123456789'
-  let length = strLength
-  while (length--) {
-    result.push(charSet.charAt(Math.floor(Math.random() * charSet.length)))
-  }
-  return result.join('')
-}
-
-/**
- * @ignore
+ * ブラウザを判定する
  */
 export function browser(): string {
   const ua = window.navigator.userAgent.toLocaleLowerCase()
@@ -49,7 +34,7 @@ export function browser(): string {
 }
 
 /**
- * @ignore
+ * デバッグログを出力する
  */
 export function traceLog(title: string, value?: string | Record<string, any>): void {
   let prefix = ''
@@ -63,76 +48,84 @@ export function traceLog(title: string, value?: string | Record<string, any>): v
   }
 }
 
-// Stack Overflow より引用: https://stackoverflow.com/a/52760103
-// https://stackoverflow.com/questions/52738290/how-to-remove-video-codecs-in-webrtc-sdp
-/** @private */
-export function getVideoCodecsFromString(codec: VideoCodecOption, codecs: Array<any>): Array<any> {
-  let mimeType = ''
-  if (codec === 'VP8') {
-    mimeType = 'video/VP8'
-  } else if (codec === 'VP9') {
-    mimeType = 'video/VP9'
-  } else if (codec === 'H264') {
-    mimeType = 'video/H264'
-  } else {
-    mimeType = `video/${codec}`
-  }
-  const filteredCodecs: Array<any> = codecs.filter((c) => c.mimeType === mimeType)
-  if (filteredCodecs.length < 1) {
-    throw new Error('invalid video codec type')
-  }
+/**
+ * 指定された codec にマッチする codec のリストを返す
+ * リストなのはプロファイルが複数合ったり、 RTX, RED, ULPFEC などの codec も含めるため
+ */
+export const getSelectedCodecs = (
+  kind: 'audio' | 'video',
+  selectedCodecMimeType: string,
+  codecs: RTCRtpCodec[],
+): RTCRtpCodecCapability[] => {
+  const filteredCodecs = codecs.filter((c) => {
+    const codecMimeType = c.mimeType.toLowerCase()
+
+    // 指定された codec はマッチしたら true
+    if (codecMimeType === selectedCodecMimeType.toLowerCase()) {
+      return true
+    }
+
+    // rtx, red, ulpfec は常に true にする
+    if (
+      codecMimeType === `${kind}/rtx` ||
+      codecMimeType === `${kind}/red` ||
+      codecMimeType === `${kind}/ulpfec`
+    ) {
+      return true
+    }
+
+    return false
+  })
   return filteredCodecs
 }
 
 /**
- * @ignore
+ * 利用可能な映像のコーデックを取得する
  */
-export function removeCodec(sdp: string, codec: VideoCodecOption): string {
-  function internalFunc(tmpSdp: string): string {
-    // eslint-disable-next-line no-useless-escape
-    const codecre = new RegExp(`(a=rtpmap:(\\d*) ${codec}/90000\\r\\n)`)
-    const rtpmaps = tmpSdp.match(codecre)
-    if (rtpmaps == null || rtpmaps.length <= 2) {
-      return sdp
-    }
-    const rtpmap = rtpmaps[2]
-    let modsdp = tmpSdp.replace(codecre, '')
-
-    const rtcpre = new RegExp(`(a=rtcp-fb:${rtpmap}.*\r\n)`, 'g')
-    modsdp = modsdp.replace(rtcpre, '')
-
-    const fmtpre = new RegExp(`(a=fmtp:${rtpmap}.*\r\n)`, 'g')
-    modsdp = modsdp.replace(fmtpre, '')
-
-    const aptpre = new RegExp(`(a=fmtp:(\\d*) apt=${rtpmap}\\r\\n)`)
-    const aptmaps = modsdp.match(aptpre)
-    let fmtpmap = ''
-    if (aptmaps != null && aptmaps.length >= 3) {
-      fmtpmap = aptmaps[2]
-      modsdp = modsdp.replace(aptpre, '')
-
-      const rtppre = new RegExp(`(a=rtpmap:${fmtpmap}.*\r\n)`, 'g')
-      modsdp = modsdp.replace(rtppre, '')
-    }
-
-    const videore = /(m=video.*\r\n)/
-    const videolines = modsdp.match(videore)
-    if (videolines != null) {
-      //If many m=video are found in SDP, this program doesn't work.
-      const videoline = videolines[0].substring(0, videolines[0].length - 2)
-      const videoelems = videoline.split(' ')
-      let modvideoline = videoelems[0]
-      videoelems.forEach((videoelem, index) => {
-        if (index === 0) return
-        if (videoelem === rtpmap || videoelem === fmtpmap) {
-          return
-        }
-        modvideoline += ` ${videoelem}`
-      })
-      modvideoline += '\r\n'
-      modsdp = modsdp.replace(videore, modvideoline)
-    }
-    return internalFunc(modsdp)
+export const getAvailableCodecs = (
+  kind: 'audio' | 'video',
+  direction: 'sender' | 'receiver',
+): string[] => {
+  if (typeof RTCRtpSender === 'undefined' || typeof RTCRtpReceiver === 'undefined') {
+    return []
   }
-  return internalFunc(sdp)
+
+  // sendrecv と sendonly は RTCRtpSender を使う
+  // recvonly は RTCRtpReceiver を使う
+  const getCapabilities =
+    direction === 'sender' || direction === 'receiver'
+      ? RTCRtpSender.getCapabilities
+      : RTCRtpReceiver.getCapabilities
+
+  if (typeof getCapabilities !== 'function') {
+    return []
+  }
+
+  const codecs = getCapabilities(kind)?.codecs
+  if (!codecs) {
+    return []
+  }
+
+  return (
+    codecs
+      .filter((c) => {
+        // mimeType は insensitive-case なので lowerCase に変換する
+        const codecType = c.mimeType.toLowerCase()
+
+        // rtx/red/ulpfec はフィルターとして削除する
+        if (
+          codecType === `${kind}/rtx` ||
+          codecType === `${kind}/red` ||
+          codecType === `${kind}/ulpfec`
+        ) {
+          return false
+        }
+
+        return true
+      })
+      // mimeType が既に存在している場合は重複を削除する
+      .filter((c, index, self) => index === self.findIndex((t) => t.mimeType === c.mimeType))
+      .map((c) => c.mimeType)
+      .sort()
+  )
 }