+++ CHANGES.md
@@ -11,14 +11,20 @@
 
 ## develop
 
+- [CHANGE] ConnectionVideoOption の codec を codecMimeType へ変更する
+  - @voluntas
 - [ADD] standalone モードに対応する
   - options に standalone を追加する
   - standalone モード時は、接続完了時に ayame に type: connected を送信する
   - standalone モード時は、ayame から WebSocket 接続が切断されても、ブラウザ間の接続は維持する
   - @Hexa
+- [FIX] disconnect の処理が正常に動作しない問題を修正する
+  - @voluntas
 
 ### misc
 
+- [CHANGE] ConnectionBase を Connection へ変更する
+  - @voluntas
 - [CHANGE] rollup から vite へ変更
   - @voluntas
 - [CHANGE] npm から pnpm に変更する
@@ -30,6 +36,8 @@
 - [CHANGE] GitHub Actions の node-version を 18 と 20 と 22 にする
   - @voluntas
 - [UPDATE] ubuntu-latest から ubuntu-24.04 に変更する
+  - @voluntas
+- [ADD] DevTools を追加
   - @voluntas
 
 ## 2022.1